pyproject.toml
[tool.poetry]
name = "database-query-builder-api"
version = "0.1.0"
description = ""
authors = ["romashorodok <108732029+romashorodok@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
django-stubs = "^5.1.0"
psycopg = {extras = ["binary", "pool"], version = "^3.2.2"}
setuptools = "^75.1.0"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
